rip out timesync and snapshot manager (#73)
"Render the newest frame you have" combined with 240 fps snapshot distribution is so superior to the old solution that it's time to delete it all.

Config parameters are preserved for the time being so that render client computers do not necessarily need to be updated to continue to be compatible.